Extract logged component
