+++ packages/client/src/modules/auth/components/LoggedUser.tsx
@@ -1,0 +1,62 @@
+import { AccountCircle } from "@mui/icons-material";
+import { IconButton, Menu, MenuItem } from "@mui/material";
+import axios from "axios";
+import React from "react";
+import { useQueryClient } from "react-query";
+
+export { LoggedUser };
+
+function LoggedUser() {
+  const queryClient = useQueryClient();
+
+  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+
+  const handleMenu = (event: React.MouseEvent<HTMLElement>) => {
+    setAnchorEl(event.currentTarget);
+  };
+
+  const handleClose = () => {
+    setAnchorEl(null);
+  };
+
+  return (
+    <div>
+      <IconButton
+        size="large"
+        aria-label="account of current user"
+        aria-controls="menu-appbar"
+        aria-haspopup="true"
+        onClick={handleMenu}
+        color="inherit"
+      >
+        <AccountCircle />
+      </IconButton>
+      <Menu
+        id="menu-appbar"
+        anchorEl={anchorEl}
+        anchorOrigin={{
+          vertical: "top",
+          horizontal: "right",
+        }}
+        keepMounted
+        transformOrigin={{
+          vertical: "top",
+          horizontal: "right",
+        }}
+        open={Boolean(anchorEl)}
+        onClose={handleClose}
+      >
+        <MenuItem
+          onClick={() => {
+            axios.post("/api/users/logout").then(() => {
+              queryClient.invalidateQueries("logged-user");
+              handleClose();
+            });
+          }}
+        >
+          Logout
+        </MenuItem>
+      </Menu>
+    </div>
+  );
+}
