+++ packages/client/src/modules/auth/components/index.ts
@@ -1,0 +1,3 @@
+import { LoggedUser } from "./LoggedUser";
+
+export { LoggedUser };
