+++ packages/client/src/modules/auth/index.ts
@@ -1,0 +1,1 @@
+export * from "./components";
